backend/core-api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.12.1
slowapi==0.1.9 